+++ docs/source/release_notes.rst
@@ -25,6 +25,7 @@ Development - |version|
 * Updated the list of `publications <publications.rst>`_
 * Added the option to compute value with sMDP rewards at the start of the step in the
   RLlib configuration.
+* Add the ability to observe remaining time in :class:`~bsk_rl.obs.Time`.
 
 
 Version 1.1.0